tests: Moved generic rest request mapper tests to separated test class.
